requirements.txt
langchain
langchain-community
openai
langchain_openai
faiss-cpu
unstructured
pdfplumber
tiktoken
python-dotenv
PyPDF2
streamlit